+++ typescript/.changeset/slimy-baths-marry.md
@@ -1,0 +1,5 @@
+---
+'@x402/evm': patch
+---
+
+Simulate transaction in verify and (optional) settle; Added multicall utility for efficient rpc calls; Fixed undeployed smart wallet handling to prevent facilitator grieving and account for implementation dependent verifyTypedData
