+++ typescript/packages/mechanisms/evm/src/exact/facilitator/eip3009.ts
@@ -4,20 +4,17 @@ import {
   SettleResponse,
   VerifyResponse,
 } from "@x402/core/types";
-import {
-  encodeFunctionData,
-  getAddress,
-  Hex,
-  isAddressEqual,
-  parseErc6492Signature,
-  parseSignature,
-} from "viem";
-import { authorizationTypes, eip3009ABI } from "../../constants";
-import { multicall, ContractCall, RawContractCall } from "../../multicall";
+import { getAddress, Hex, isAddressEqual, parseErc6492Signature } from "viem";
+import { authorizationTypes } from "../../constants";
 import { FacilitatorEvmSigner } from "../../signer";
 import { getEvmChainId } from "../../utils";
 import { ExactEIP3009Payload } from "../../types";
 import * as Errors from "./errors";
+import {
+  diagnoseEip3009SimulationFailure,
+  executeTransferWithAuthorization,
+  simulateEip3009Transfer,
+} from "./eip3009-utils";
 
 export interface VerifyEIP3009Options {
   /** Run onchain simulation. Defaults to true. */
@@ -209,78 +206,20 @@ export async function verifyEIP3009(
 
   // Transaction simulation
   if (options?.simulate !== false) {
-    const auth = eip3009Payload.authorization;
-    const transferArgs = [
-      getAddress(auth.from),
-      getAddress(auth.to),
-      BigInt(auth.value),
-      BigInt(auth.validAfter),
-      BigInt(auth.validBefore),
-      auth.nonce,
-    ] as const;
-
-    let simulationFailed = false;
-
-    if (eip6492Deployment) {
-      // Undeployed EIP-6492: simulate via multicall (deploy + transfer in one eth_call)
-      const { signature: innerSignature } = parseErc6492Signature(eip3009Payload.signature!);
-      const transferCalldata = encodeFunctionData({
-        abi: eip3009ABI,
-        functionName: "transferWithAuthorization",
-        args: [...transferArgs, innerSignature],
-      });
-
-      const results = await multicall(signer.readContract.bind(signer), [
-        {
-          address: getAddress(eip6492Deployment.factoryAddress),
-          callData: eip6492Deployment.factoryCalldata as Hex,
-        } satisfies RawContractCall,
-        {
-          address: erc20Address,
-          callData: transferCalldata,
-        } satisfies RawContractCall,
-      ]);
-
-      // If the transfer call fails, the simulation failed
-      // Smart wallet deployment might succeed even if the inner signature is invalid, but the transfer will fail
-      if (results[1]?.status !== "success") {
-        simulationFailed = true;
-      }
-    } else {
-      // EOA or deployed smart wallet
-      const sig = eip3009Payload.signature!;
-      const sigLength = sig.startsWith("0x") ? sig.length - 2 : sig.length;
-      const isECDSA = sigLength === 130;
-
-      try {
-        if (isECDSA) {
-          const parsedSig = parseSignature(sig);
-          await signer.readContract({
-            address: erc20Address,
-            abi: eip3009ABI,
-            functionName: "transferWithAuthorization",
-            args: [
-              ...transferArgs,
-              (parsedSig.v as number | undefined) ?? parsedSig.yParity,
-              parsedSig.r,
-              parsedSig.s,
-            ],
-          });
-        } else {
-          await signer.readContract({
-            address: erc20Address,
-            abi: eip3009ABI,
-            functionName: "transferWithAuthorization",
-            args: [...transferArgs, sig],
-          });
-        }
-      } catch {
-        simulationFailed = true;
-      }
-    }
-
+    const simulationFailed = await simulateEip3009Transfer(
+      signer,
+      erc20Address,
+      eip3009Payload,
+      eip6492Deployment,
+    );
     if (simulationFailed) {
-      return diagnoseSimulationFailure(signer, erc20Address, eip3009Payload, requirements);
+      return diagnoseEip3009SimulationFailure(
+        signer,
+        erc20Address,
+        eip3009Payload,
+        requirements,
+        requirements.amount,
+      );
     }
   }
 
@@ -289,96 +228,6 @@ export async function verifyEIP3009(
     invalidReason: undefined,
     payer,
   };
-}
-
-/**
- * After simulation fails, runs a single diagnostic multicall to determine the most specific error reason.
- * Checks balanceOf, name, version and authorizationState in one RPC round-trip.
- *
- * @param signer - EVM signer used for the payment
- * @param erc20Address - Address of the ERC-20 token contract
- * @param eip3009Payload - The EIP-3009 transfer authorization payload
- * @param requirements - Payment requirements to validate against
- * @returns A promise that resolves to the verification result with validity and optional invalid reason
- */
-async function diagnoseSimulationFailure(
-  signer: FacilitatorEvmSigner,
-  erc20Address: `0x${string}`,
-  eip3009Payload: ExactEIP3009Payload,
-  requirements: PaymentRequirements,
-): Promise<VerifyResponse> {
-  const payer = eip3009Payload.authorization.from;
-
-  const diagnosticCalls: ContractCall[] = [
-    {
-      address: erc20Address,
-      abi: eip3009ABI,
-      functionName: "balanceOf",
-      args: [eip3009Payload.authorization.from],
-    },
-    {
-      address: erc20Address,
-      abi: eip3009ABI,
-      functionName: "name",
-    },
-    {
-      address: erc20Address,
-      abi: eip3009ABI,
-      functionName: "version",
-    },
-    {
-      address: erc20Address,
-      abi: eip3009ABI,
-      functionName: "authorizationState",
-      args: [eip3009Payload.authorization.from, eip3009Payload.authorization.nonce],
-    },
-  ];
-
-  try {
-    const results = await multicall(signer.readContract.bind(signer), diagnosticCalls);
-
-    const [balanceResult, nameResult, versionResult, authStateResult] = results;
-
-    // authorizationState reverted -> token doesn't support EIP-3009
-    if (authStateResult.status === "failure") {
-      return { isValid: false, invalidReason: Errors.ErrEip3009NotSupported, payer };
-    }
-
-    // authorizationState == true -> nonce already used
-    if (authStateResult.status === "success" && authStateResult.result === true) {
-      return { isValid: false, invalidReason: Errors.ErrEip3009NonceAlreadyUsed, payer };
-    }
-
-    // name mismatch
-    if (
-      nameResult.status === "success" &&
-      requirements.extra?.name &&
-      nameResult.result !== requirements.extra.name
-    ) {
-      return { isValid: false, invalidReason: Errors.ErrEip3009TokenNameMismatch, payer };
-    }
-
-    // version mismatch
-    if (
-      versionResult.status === "success" &&
-      requirements.extra?.version &&
-      versionResult.result !== requirements.extra.version
-    ) {
-      return { isValid: false, invalidReason: Errors.ErrEip3009TokenVersionMismatch, payer };
-    }
-
-    // insufficient balance
-    if (balanceResult.status === "success") {
-      const balance = balanceResult.result as bigint;
-      if (balance < BigInt(requirements.amount)) {
-        return { isValid: false, invalidReason: Errors.ErrEip3009InsufficientBalance, payer };
-      }
-    }
-  } catch {
-    // Diagnostic multicall failed — fall through to generic error
-  }
-
-  return { isValid: false, invalidReason: Errors.ErrEip3009SimulationFailed, payer };
 }
 
 /**
@@ -415,9 +264,10 @@ export async function settleEIP3009(
   }
 
   try {
-    // Parse ERC-6492 signature if applicable
-    const parseResult = parseErc6492Signature(eip3009Payload.signature!);
-    const { signature, address: factoryAddress, data: factoryCalldata } = parseResult;
+    // Parse ERC-6492 signature if applicable (for optional deployment)
+    const { address: factoryAddress, data: factoryCalldata } = parseErc6492Signature(
+      eip3009Payload.signature!,
+    );
 
     // Deploy ERC-4337 smart wallet via EIP-6492 if configured and needed
     if (
@@ -441,48 +291,11 @@ export async function settleEIP3009(
       }
     }
 
-    // Determine if this is an ECDSA signature (EOA) or smart wallet signature
-    const signatureLength = signature.startsWith("0x") ? signature.length - 2 : signature.length;
-    const isECDSA = signatureLength === 130;
-
-    let tx: Hex;
-    if (isECDSA) {
-      // For EOA wallets, parse signature into v, r, s and use that overload
-      const parsedSig = parseSignature(signature);
-
-      tx = await signer.writeContract({
-        address: getAddress(requirements.asset),
-        abi: eip3009ABI,
-        functionName: "transferWithAuthorization",
-        args: [
-          getAddress(eip3009Payload.authorization.from),
-          getAddress(eip3009Payload.authorization.to),
-          BigInt(eip3009Payload.authorization.value),
-          BigInt(eip3009Payload.authorization.validAfter),
-          BigInt(eip3009Payload.authorization.validBefore),
-          eip3009Payload.authorization.nonce,
-          (parsedSig.v as number | undefined) || parsedSig.yParity,
-          parsedSig.r,
-          parsedSig.s,
-        ],
-      });
-    } else {
-      // For smart wallets, use the bytes signature overload
-      tx = await signer.writeContract({
-        address: getAddress(requirements.asset),
-        abi: eip3009ABI,
-        functionName: "transferWithAuthorization",
-        args: [
-          getAddress(eip3009Payload.authorization.from),
-          getAddress(eip3009Payload.authorization.to),
-          BigInt(eip3009Payload.authorization.value),
-          BigInt(eip3009Payload.authorization.validAfter),
-          BigInt(eip3009Payload.authorization.validBefore),
-          eip3009Payload.authorization.nonce,
-          signature,
-        ],
-      });
-    }
+    const tx = await executeTransferWithAuthorization(
+      signer,
+      getAddress(requirements.asset),
+      eip3009Payload,
+    );
 
     // Wait for transaction confirmation
     const receipt = await signer.waitForTransactionReceipt({ hash: tx });
